refactor(catalog): apply refresh-impl review feedback
- Add CatalogError.network case (HTTP errors no longer mislabelled as .sqlite)
- Extract LockedBox to MoolahTests/Support/ (deduplicate across two test files)
- Correct cross-extension internals MARK to attribute helpers per file
- Replace URL '?? URL(fileURLWithPath:)' fallback with guarded preconditionFailure
- Replace HTTPURLResponse '?? HTTPURLResponse()' with force-unwrap (test-only) + safe comment
- Add lastFetched-updated assertion to the 304 test (proves design invariant)
- Mark refresh + search-client suites .serialized (StubURLProtocol.handlers race)
- Conform CoinsUpdate / PlatformsUpdate to Sendable (matches project pattern)
- Use async let for parallel /coins/list and /asset_platforms fetches (per plan §5.4)
